requirements.txt
Flask
psutil
flask-socketio
eventlet
pywebview